@calmh 2021-10-29
Yeah deepsource can go pound sand in this case.